Detect time complexity in Java

+++ java/Example_1.java
@@ -1,7 +1,7 @@
 public class Example_1 {
     public static void start(long n) {
         long c = 0;
-        for (int i = 0; i < n; i++) {
+        for (long i = 0; i < n; i++) {
             c++;
         }
     }
